Don't Allow User to follow/unfollow himself

+++ app/controllers/relationships_controller.rb
@@ -3,7 +3,7 @@ class RelationshipsController < ApplicationController
 
 	def follow
 		@user = User.where(:username => params[:username]).first
-		if not @user.nil?
+		if not @user.nil? and current_user!=@user
 			@user.followers << current_user
 			@user.save!
 		end
@@ -14,7 +14,7 @@ class RelationshipsController < ApplicationController
 
 	def unfollow
 		@user = User.where(:username => params[:username]).first
-		if not @user.nil?
+		if not @user.nil? and current_user!=@user
 			relation = Relationship.where(:follower_id => current_user.id, :following_id => @user.id).first
 			relation.destroy
 		end
